wip: Let user click through organizations with animations.
Instead of listing out all organizations, let the user
click through to the next organization and view each
card of data at a time.

Disable buttons when there are no more organizations
to go to (next and previous).

Add basic fade animations when the app loads (see previous commit)
and when a new organization is selected. To do this without an abrupt
effect on the user I had to wrap the container with another div
that has the same background colour as the animated/faded sections.
Without this it would flash white at first. If I put the animation
container within the section flex container it would insert a span
and break my flexbox container.