Add data for organizations.
Add some real data for each ministry.

This data was manually calculated, the 
data structure should be changed and the code base 
should be updated to take the raw data available
and calculate the proper values as needed. 

These numbers may be incorrect due to human error, 
they are also a count instead of a percentage like
the charts are expecting.